cleaner mouse click coord copy implementation
